% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adtseq}
\alias{adtseq}
\title{Identify antibody-derived-tags from single cell RNAseq data}
\usage{
adtseq(bamFileName, bamOut, adtFasta, max_dist, sumoutput)
}
\arguments{
\item{bamFileName}{The input BAM/SAM filepath, this is typically output from the dropseq-tools pipeline.}

\item{bamOut}{The output BAM/SAM filepath}

\item{adtFasta}{Filepath for the fasta file containing the TotalSeq antibody barcode sequences (in fasta format)}

\item{max_dist}{integer. The maximum Hamming distance allowed for a barcode match to be called}

\item{sumoutput}{Filepath for tab delimited text file output of cell-barcodes, umi, antibody, hamming distance and map score for each record in the output bam/sam}
}
\description{
Identify antibody-derived-tags from single cell RNAseq data
}
