% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bcExtract}
\alias{bcExtract}
\title{Extraction of read names from specific cell barcodes.}
\usage{
bcExtract(bamIn, bamOut, sumoutput, cellBC)
}
\arguments{
\item{bamIn}{The input BAM/SAM filepath, this is typically output from the dropseq-tools pipeline (string).}

\item{bamOut}{The output BAM/SAM filepath (string).}

\item{sumoutput}{Filepath for tab delimited text file output of cell-barcodes, umi, fastq read name (string).}

\item{cellBC}{Character string specifying the 12 nt cell barcode for which you want to extract reads for.}
}
\description{
Extraction of read names from specific cell barcodes.
}
