% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexityEntropy.R
\name{complexityEntropy}
\alias{complexityEntropy}
\title{Entropic Complexity of Sequences}
\usage{
complexityEntropy(object)
}
\arguments{
\item{object}{A DNAStringSet}
}
\description{
Takes a DNAStringSet and calculates the entropic complexity of each sequence within the set.
}
\author{
Julian Spagnuolo
}
