% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hdist}
\alias{hdist}
\title{Pairwise Hamming Distance}
\usage{
hdist(seqs)
}
\arguments{
\item{seqs}{Character vector of equal length strings containing the sequences from which to calculate pairwise Hamming/String distances.}
}
\description{
Pairwise Hamming Distance
}
