% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icgrEncode}
\alias{icgrEncode}
\title{Integer Chaos Game Encoder}
\arguments{
\item{seq, }{a character vector of length = 1}
}
\description{
This function will encode a DNA sequence as an integer chaos game representation see: 
[Yin, C. (2018). Encoding DNA sequences by integer Chaos Game Representation, Journal of Computational Biology](https://arxiv.org/abs/1712.04546)
for more detail.
}
\author{
Julian Spagnuolo
}
