% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/package.R
\docType{package}
\name{adtseq}
\alias{adtseq}
\alias{adtseq-package}
\title{Antibody Derived Tag Extraction}
\arguments{
\item{bamFileName, }{An processed BAM or SAM file that has been tagged with the cell barcode and UMI using the Dropseq-tools TagBamWithReadSequenceExtended.}

\item{bamOut, }{output BAM/SAM file, will autodetect format by the suffix i.e. *.bam for BAM, *.sam for SAM}

\item{adtFasta, }{a fasta file containing the sequences for the antibody derived tags}

\item{max_dist, }{Integer, the maximum allowed hamming distance a read can be from the expected antibody derived tag sequence in the fasta file}

\item{sumoutput, }{output summary txt file}
}
\description{
This function will read a processed/tagged DropSeq (not tested on 10x) BAM/SAM file, tagging each record with antibody barcode name (from the provided ADT-tag fasta file)
using a combination of hamming distance (less than max_dist) and best Phred quality weighted mapping score calculated as the sum of scores for each base, 
subtracting the score at mismatches or adding at matches. This is performed similarly to the method used by Bowtie2 to determine alignment scores.
At present Bowtie2’s defaults for MN (minimum score) and MX (maximum score), 2 and 6, respectively, are used.
The function will place the name of the identified antibody in a new tag “XA”, the mapping score under “AS”, and the hamming distance under “XN”.
Additionally, the function will output a second tab-delimited text file containing the Cell Barcodes, UMI and identified antibody along with the hamming distance and mapping scores for each bam/sam record.
The resulting BAM/SAM file can be directly used in latter stages of the dropseq-tools pipeline which attempt to correct for synthesis errors in the cell barcodes and UMIs.
However, should users wish, they may retrieve count data directly from the text file and omit the latter steps (not recommended).

Identification of Antibody-Derived-Tags in single cell CITE/REAP-seq data
}
\author{
Julian Spagnuolo

Julian Spagnuolo
}
