% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bcExtract}
\alias{bcExtract}
\title{Barcode Extractor}
\usage{
bcExtract(bamIn, bamOut, sumoutput, cellBC)
}
\arguments{
\item{bamIn, }{A processed BAM or SAM file that has been tagged with the cell barcode and UMI using the Dropseq-tools TagBamWithReadSequenceExtended.}

\item{bamOut, }{output BAM/SAM file, will autodetect format by the suffix.}

\item{sumoutput, }{output summary txt file}

\item{cellBC, }{deprecated and unused.}
}
\description{
This function will read a processed/tagged DropSeq BAM/SAM file, extract cell barcodes and UMIs from the BAM/SAM tags,
and count the nucleotide frequency within the barcode, UMI and the read itself. NOTE: not tested on 10x
It's intended function is to provide information on the diversity of the sequences within the DropSeq reads.
}
\author{
Julian Spagnuolo
}
