% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hdist}
\alias{hdist}
\title{Hamming Distance}
\usage{
hdist(umi)
}
\arguments{
\item{umi, }{character vector of strings.}
}
\description{
Simple function to determine the Hamming or String distance between each pair of strings in a vector. Outputs an integer matrix.
}
\author{
Julian Spagnuolo
}
